Change things about a bit for what FuzzyLockKeys default to for the non fuzzy type
