+++ src/api/java/com/teamacronymcoders/eposmajorum/api/locks/IGenericFuzzyLockType.java
@@ -1,0 +1,4 @@
+package com.teamacronymcoders.eposmajorum.api.locks;
+
+public interface IGenericFuzzyLockType {
+}
