+++ src/main/java/com/teamacronymcoders/eposmajorum/locks/keys/ArmorToughnessLockKey.java
@@ -1,7 +1,9 @@
 package com.teamacronymcoders.eposmajorum.locks.keys;
 
+import com.teamacronymcoders.eposmajorum.api.locks.GenericLockKey;
 import com.teamacronymcoders.eposmajorum.api.locks.IFuzzyLockKey;
 import com.teamacronymcoders.eposmajorum.api.locks.ILockKey;
+import com.teamacronymcoders.eposmajorum.locks.FuzzyLockKeyTypes;
 import java.util.Collection;
 import java.util.Objects;
 import javax.annotation.Nonnull;
@@ -53,9 +55,9 @@ public class ArmorToughnessLockKey implements IFuzzyLockKey {
     }
 
     @Override
-    @Nullable
+    @Nonnull
     public ILockKey getNotFuzzy() {
-        return null;
+        return new GenericLockKey(FuzzyLockKeyTypes.ARMOR_TOUGHNESS);
     }
 
     @Override
