+++ src/main/java/com/teamacronymcoders/eposmajorum/locks/keys/AttackDamageLockKey.java
@@ -1,8 +1,10 @@
 package com.teamacronymcoders.eposmajorum.locks.keys;
 
 import com.google.common.collect.Multimap;
+import com.teamacronymcoders.eposmajorum.api.locks.GenericLockKey;
 import com.teamacronymcoders.eposmajorum.api.locks.IFuzzyLockKey;
 import com.teamacronymcoders.eposmajorum.api.locks.ILockKey;
+import com.teamacronymcoders.eposmajorum.locks.FuzzyLockKeyTypes;
 import java.util.Collection;
 import java.util.Objects;
 import javax.annotation.Nonnull;
@@ -56,9 +58,9 @@ public class AttackDamageLockKey implements IFuzzyLockKey {
     }
 
     @Override
-    @Nullable
+    @Nonnull
     public ILockKey getNotFuzzy() {
-        return null;
+        return new GenericLockKey(FuzzyLockKeyTypes.ATTACK_DAMAGE);
     }
 
     @Override
